refact(research): extract CaptchaNotification inline fetching to useCaptchaStatus (#6082)
